Finish The Home Screen Design and Functinalities
Implement the navigation routes for the Home Screen and it functinalities; Upon clicking a category one is navigated to a list of the categories. upon clicking the 'see more' Text Button a user is navigated to the list of all the top bikes, on clicking a bike a user is navigated to the Bike Details Screen.

Fix the fetching of bikes in a "not intended" way. Now all the bikes that "EXCELLENT" are shown first then the ones that are "GOOD"